814. Binary Tree Pruning
https://leetcode.com/problems/binary-tree-pruning/
